[eto] improve codegen
