+++ src/skiasharp/LiveChartsCore.SkiaSharpView.Eto/PieChart.cs
@@ -31,6 +31,13 @@ using LiveChartsCore.Kernel.Sketches;
 
 namespace LiveChartsCore.SkiaSharpView.Eto;
 
+// ==============================================================================
+// 
+// use the LiveChartsGeneratedCode.PieChart class to add eto specific
+// code, this class is just to expose the PieChart class in this namespace.
+// 
+// ==============================================================================
+
 /// <inheritdoc cref="IPieChartView" />
-public partial class PieChart : ChartControl, IPieChartView
+public partial class PieChart : LiveChartsGeneratedCode.PieChart
 { }
